实时数据改用dataframe-go存储

+++ common/stockOptions.go
@@ -11,7 +11,7 @@ type RankOpt struct {
 	MarketType string // 市场类型
 	SortName   string
 	Sorted     bool //排序
-	Page       int64
+	Page       int
 }
 
 // SearchOpt 搜索
